+++ RechargeSharp.v2021-11/Endpoints/Customers/CustomerService.Endpoints.cs
@@ -7,11 +7,11 @@ namespace RechargeSharp.v2021_11.Endpoints.Customers;
 public interface ICustomerService
 {
     Task<CustomerService.CreateCustomerTypes.Response> CreateCustomerAsync(CustomerService.CreateCustomerTypes.Request request);
-    Task<CustomerService.GetCustomerTypes.Response?> GetCustomerAsync(int customerId);
-    Task<CustomerService.UpdateCustomerTypes.Response> UpdateCustomerAsync(int customerId, CustomerService.UpdateCustomerTypes.Request request);
-    Task DeleteCustomerAsync(int customerId);
+    Task<CustomerService.GetCustomerTypes.Response?> GetCustomerAsync(long customerId);
+    Task<CustomerService.UpdateCustomerTypes.Response> UpdateCustomerAsync(long customerId, CustomerService.UpdateCustomerTypes.Request request);
+    Task DeleteCustomerAsync(long customerId);
     Task<CustomerService.ListCustomersTypes.Response> ListCustomersAsync(CustomerService.ListCustomersTypes.Request? request);
-    Task<CustomerService.GetCustomerDeliveryScheduleTypes.Response> GetCustomerDeliveryScheduleAsync(int customerId, CustomerService.GetCustomerDeliveryScheduleTypes.Request request);
+    Task<CustomerService.GetCustomerDeliveryScheduleTypes.Response> GetCustomerDeliveryScheduleAsync(long customerId, CustomerService.GetCustomerDeliveryScheduleTypes.Request request);
 }
 
 public partial class CustomerService : ICustomerService
@@ -30,21 +30,22 @@ public partial class CustomerService : ICustomerService
         return responseJson;
     }
 
-    public async Task<GetCustomerTypes.Response?> GetCustomerAsync(int customerId)
+    public async Task<GetCustomerTypes.Response?> GetCustomerAsync(long customerId)
     {
         var requestUri = $"/customers/{customerId}";
         var responseJson = await _rechargeApiCaller.GetNullableAsync<GetCustomerTypes.Response>(requestUri);
         return responseJson;
     }
     
-    public async Task<UpdateCustomerTypes.Response> UpdateCustomerAsync(int customerId, UpdateCustomerTypes.Request request)
+    public async Task<UpdateCustomerTypes.Response> UpdateCustomerAsync(long customerId,
+        UpdateCustomerTypes.Request request)
     {
         var requestUri = $"/customers/{customerId}";
         var responseJson = await _rechargeApiCaller.PutAsync<UpdateCustomerTypes.Request, UpdateCustomerTypes.Response>(request, requestUri);
         return responseJson;
     }
     
-    public async Task DeleteCustomerAsync(int customerId)
+    public async Task DeleteCustomerAsync(long customerId)
     {
         var requestUri = $"/customers/{customerId}";
         await _rechargeApiCaller.DeleteAsync(requestUri);
@@ -58,7 +59,8 @@ public partial class CustomerService : ICustomerService
         return responseJson;
     }
     
-    public async Task<GetCustomerDeliveryScheduleTypes.Response> GetCustomerDeliveryScheduleAsync(int customerId, GetCustomerDeliveryScheduleTypes.Request request)
+    public async Task<GetCustomerDeliveryScheduleTypes.Response> GetCustomerDeliveryScheduleAsync(long customerId,
+        GetCustomerDeliveryScheduleTypes.Request request)
     {
         var queryString = ObjectToQueryStringSerializer.SerializeObjectToQueryString(request);
         var requestUri = $"/customers/{customerId}/delivery_schedule{queryString}";
